backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
django-cors-headers==3.2.1
django-environ==0.4.5
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
environ==1.0
gunicorn==20.0.4
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
pydantic==1.8.1
python-dateutil==2.8.1
PyJWT==1.7.1
python-environ==0.4.54
pytz==2021.1
redis==3.5.3
regex==2021.3.17
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
